-fixed compat with Deeeper&Darker -moved compat classes into items/compats
