+++ build.gradle
@@ -66,7 +66,7 @@ dependencies {
 	modCompileOnly "maven.modrinth:deeperdarker:${deeperanddarker_version}"
 	modRuntimeOnly "maven.modrinth:deeperdarker:${deeperanddarker_version}"
 	modImplementation "dev.isxander:yet-another-config-lib:${project.yacl_version}+${minecraft_version}-fabric"
-	include modImplementation("net.kyrptonaught:customportalapi:${project.customportalapi_version}")
+	modImplementation("net.kyrptonaught:customportalapi:${project.customportalapi_version}")
 	// better end
 //	modCompileOnly "maven.modrinth:betterend:${project.betterend_version}"
 //	modRuntimeOnly "maven.modrinth:betterend:${project.betterend_version}"
