+++ src/main/java/net/hayden/distinctshields/datagen/ModRecipeGenerator.java
@@ -3,8 +3,8 @@ package net.hayden.distinctshields.datagen;
 import com.kyanite.deeperdarker.content.DDBlocks;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
-import net.hayden.distinctshields.items.DeeperDarkerItems;
-import net.hayden.distinctshields.items.NaturesSpiritItems;
+import net.hayden.distinctshields.items.compats.DeeperDarkerItems;
+import net.hayden.distinctshields.items.compats.NaturesSpiritItems;
 import net.hayden.distinctshields.items.VanillaItems;
 import net.hibiscus.naturespirit.registration.block_registration.HibiscusWoods;
 import net.minecraft.data.server.recipe.RecipeJsonProvider;
