+++ src/main/java/net/hayden/distinctshields/items/compats/BetterEndItems.java
@@ -1,4 +1,4 @@
-package net.hayden.distinctshields.items;
+package net.hayden.distinctshields.items.compats;
 
 import com.google.common.collect.Maps;
 import net.hayden.distinctshields.DistinctShields;
