+++ src/main/java/net/hayden/distinctshields/util/ModAtlasesGen.java
@@ -7,7 +7,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import net.hayden.distinctshields.DistinctShields;
 import net.hayden.distinctshields.items.VanillaItems;
-import net.hayden.distinctshields.items.NaturesSpiritItems;
+import net.hayden.distinctshields.items.compats.NaturesSpiritItems;
 import net.minecraft.item.Item;
 
 import java.io.FileWriter;
